pyproject.toml
[project]
name = "whatsapp-toolkit"
version = "1.9.1"
description = "Una biblioteca para interactuar con la API de WhatsApp desde Python usando Evolution API"
authors = [
    { name="Fernando Leon Franco", email="fernanlee2131@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "colorstreak>=2.1.3",
    "fastapi>=0.124.4",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
    "piper-tts>=1.3.0",
    "platformdirs>=4.5.1",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "reportlab>=4.4.6",
    "requests>=2.32.5",
    "typer>=0.21.0",
    "uvicorn>=0.38.0",
]

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/epok200/whatsapp_toolkit"
Repository = "https://github.com/epok200/whatsapp_toolkit"
Issues = "https://github.com/epok200/whatsapp_toolkit/issues"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"


[project.scripts]
wtk = "whatsapp_toolkit.cli.main:app"
